+++ lib/src/xml/mixins/has_children.dart
@@ -16,8 +16,14 @@ mixin XmlChildrenBase {
   /// Return the first child of this node, or `null` if there are no children.
   XmlNode get firstChild => null;
 
+  /// Return the first child [XmlElement], or `null` if there are none.
+  XmlElement get firstElementChild => null;
+
   /// Return the last child of this node, or `null` if there are no children.
   XmlNode get lastChild => null;
+
+  /// Return the last child [XmlElement], or `null` if there are none.
+  XmlElement get lastElementChild => null;
 }
 
 /// Mixin for nodes with children.
@@ -34,5 +40,13 @@ mixin XmlHasChildren implements XmlChildrenBase {
   XmlNode get firstChild => children.isEmpty ? null : children.first;
 
   @override
+  XmlElement get firstElementChild =>
+      children.firstWhere((node) => node is XmlElement, orElse: () => null);
+
+  @override
   XmlNode get lastChild => children.isEmpty ? null : children.last;
+
+  @override
+  XmlElement get lastElementChild =>
+      children.lastWhere((node) => node is XmlElement, orElse: () => null);
 }